move cLog to utils.js
